JDK-8311870-Split-CompressedKlassPointers-from-compressedOops.hpp
