+++ src/hotspot/share/gc/x/xDebug.gdb
@@ -51,7 +51,7 @@ define zpo
     printf "\t Page: %llu\n", ((uintptr_t)$obj & XAddressOffsetMask) >> XGranuleSizeShift
     x/16gx $obj
     if (UseCompressedClassPointers)
-        set $klass = (Klass*)(void*)((uintptr_t)CompressedKlassPointers::_narrow_klass._base +((uintptr_t)$obj->_metadata->_compressed_klass << CompressedKlassPointers::_narrow_klass._shift))
+        set $klass = (Klass*)(void*)((uintptr_t)CompressedKlassPointers::_base +((uintptr_t)$obj->_metadata->_compressed_klass << CompressedKlassPointers::_shift))
     else
         set $klass = $obj->_metadata->_klass
     end
