+++ src/hotspot/share/oops/compressedOops.hpp
@@ -40,7 +40,7 @@ struct NarrowPtrStruct {
   address _base;
   // Number of shift bits for encoding/decoding narrow ptrs.
   // 0 if using wide ptrs or zero based unscaled narrow ptrs,
-  // LogMinObjAlignmentInBytes/LogKlassAlignmentInBytes otherwise.
+  // LogMinObjAlignmentInBytes otherwise.
   int     _shift;
   // Generate code with implicit null checks for narrow ptrs.
   bool    _use_implicit_null_checks;
@@ -138,62 +138,6 @@ public:
 
   template<typename T>
   static inline narrowOop narrow_oop_cast(T i);
-};
-
-// For UseCompressedClassPointers.
-class CompressedKlassPointers : public AllStatic {
-  friend class VMStructs;
-
-  static NarrowPtrStruct _narrow_klass;
-
-  // Together with base, this defines the address range within which Klass
-  //  structures will be located: [base, base+range). While the maximal
-  //  possible encoding range is 4|32G for shift 0|3, if we know beforehand
-  //  the expected range of Klass* pointers will be smaller, a platform
-  //  could use this info to optimize encoding.
-  static size_t _range;
-
-  static void set_base(address base);
-  static void set_range(size_t range);
-  static void set_shift(int shift);
-
-public:
-
-  // Given an address p, return true if p can be used as an encoding base.
-  //  (Some platforms have restrictions of what constitutes a valid base
-  //   address).
-  static bool is_valid_base(address p);
-
-  // Given a klass range [addr, addr+len) and a given encoding scheme, assert that this scheme covers the range, then
-  // set this encoding scheme. Used by CDS at runtime to re-instate the scheme used to pre-compute klass ids for
-  // archived heap objects.
-  static void initialize_for_given_encoding(address addr, size_t len, address requested_base, int requested_shift);
-
-  // Given an address range [addr, addr+len) which the encoding is supposed to
-  //  cover, choose base, shift and range.
-  //  The address range is the expected range of uncompressed Klass pointers we
-  //  will encounter (and the implicit promise that there will be no Klass
-  //  structures outside this range).
-  static void initialize(address addr, size_t len);
-
-  static void     print_mode(outputStream* st);
-
-  static address  base()               { return  _narrow_klass._base; }
-  static size_t   range()              { return  _range; }
-  static int      shift()              { return  _narrow_klass._shift; }
-
-  static bool is_null(Klass* v)      { return v == nullptr; }
-  static bool is_null(narrowKlass v) { return v == 0; }
-
-  static inline Klass* decode_raw(narrowKlass v, address base, int shift);
-  static inline Klass* decode_raw(narrowKlass v);
-  static inline Klass* decode_not_null(narrowKlass v);
-  static inline Klass* decode_not_null(narrowKlass v, address base, int shift);
-  static inline Klass* decode(narrowKlass v);
-  static inline narrowKlass encode_not_null(Klass* v);
-  static inline narrowKlass encode_not_null(Klass* v, address base, int shift);
-  static inline narrowKlass encode(Klass* v);
-
 };
 
 #endif // SHARE_OOPS_COMPRESSEDOOPS_HPP
