+++ src/hotspot/share/oops/oopsHierarchy.hpp
@@ -37,9 +37,6 @@
 // Global offset instead of address for an oop within a java object.
 enum class narrowOop : uint32_t { null = 0 };
 
-// If compressed klass pointers then use narrowKlass.
-typedef juint  narrowKlass;
-
 typedef void* OopOrNarrowOopStar;
 
 #ifndef CHECK_UNHANDLED_OOPS
