+++ src/hotspot/share/runtime/vmStructs.cpp
@@ -379,8 +379,8 @@
   /* CompressedKlassPointers */                                                                                                      \
   /***************************/                                                                                                      \
                                                                                                                                      \
-     static_field(CompressedKlassPointers,     _narrow_klass._base,                           address)                               \
-     static_field(CompressedKlassPointers,     _narrow_klass._shift,                          int)                                   \
+     static_field(CompressedKlassPointers,     _base,                                         address)                               \
+     static_field(CompressedKlassPointers,     _shift,                                        int)                                   \
                                                                                                                                      \
   /**********/                                                                                                                       \
   /* Memory */                                                                                                                       \
