+++ src/jdk.hotspot.agent/share/classes/sun/jvm/hotspot/oops/CompressedKlassPointers.java
@@ -58,8 +58,8 @@ public class CompressedKlassPointers {
   private static synchronized void initialize(TypeDataBase db) {
     Type type = db.lookupType("CompressedKlassPointers");
 
-    baseField = type.getAddressField("_narrow_klass._base");
-    shiftField = type.getCIntegerField("_narrow_klass._shift");
+    baseField = type.getAddressField("_base");
+    shiftField = type.getCIntegerField("_shift");
   }
 
   public CompressedKlassPointers() {
